\name{Data_Jena_20_metacommunities}
\alias{Data_Jena_20_metacommunities}
\docType{data}
\title{
Biomass time series data of 20 sets of plots in the Jena experiment
}
\description{
  All 76 plots are grouped into 20 metacommunities according to block identity and species-richness level. Each metacommunity consists of three or four plots (communities). All plots within a metacommunity have the same species richness but differ in species composition. This dataset includes the 22-year (2003 to 2024) biomass time series for constituent communities within each metacommunity.\cr
}

\usage{data(Data_Jena_20_metacommunities)}
\format{
  \code{Data_Jena_20_metacommunities} contains 20 lists; each list is a data frame with plot identifiers as rows and time points as columns.
 }


\keyword{datasets}
