\name{Data_Jena_462_populations}
\alias{Data_Jena_462_populations}
\docType{data}
\title{
Biomass time series data of 462 populations in the Jena experiment
}
\description{
  The Jena Experiment consists of 76 plots arranged into 4 blocks, each containing 18–20 plots. Plots were sown along a gradient of 1, 2, 4, 8, or 16 species, resulting in a total of 462 populations. The biomass of each species in each plot was recorded annually from 2003 to 2024, except in 2004. This dataset contains the 21-year biomass time series for all 462 populations across 76 plots. This dataset is a population-by-time (462 × 21) data frame, in which each row represents a population and each column represents one year of biomass data.\cr
}
\usage{data(Data_Jena_462_populations)}
\format{
  \code{Data_Jena_462_populations} is a population-by-time data frame with 462 rows and 21 columns. Each row represents a population, and each column represents one year of biomass data.
 }

\keyword{datasets}
