\name{Data_Jena_76_communities}
\alias{Data_Jena_76_communities}
\docType{data}
\title{
Biomass time series data of 76 communities in the Jena experiment
}
\description{
  The biomass data for all species sown within each plot (community) are used to form 76 community datasets. Because the number of species sown per plot ranges from 1 to 16, the number of populations within a community also ranges from 1 to 16. Each community dataset contains the 21-year biomass time series for all constituent species-level populations. \cr
}

\usage{data(Data_Jena_76_communities)}
\format{
  \code{Data_Jena_76_communities} is a list of 76 data frames; each data frame has population identifiers as rows and time points as columns.
 }

\keyword{datasets}
