\name{Data_Jena_76_community_populations}
\alias{Data_Jena_76_community_populations}
\docType{data}
\title{
Biomass time series data of 76 communities in the Jena experiment
}
\description{
  The biomass data for all species sown within each plot are used to form 76 community datasets. Because the number of species sown per plot ranges from 1 to 16, the number of populations within a community also ranges from 1 to 16. Each community dataset contains the 21-year biomass time series for all constituent populations. The content of this dataset is identical to that of `Data_Jena_462_populations`, but for analytical convenience, it is reorganized into a list of 76 communities, each containing a population-by-time biomass data frame for its constituent populations.\cr
}

\usage{data(Data_Jena_76_community_populations)}
\format{
  \code{Data_Jena_76_community_populations} is a list of 76 data frames; each data frame has population identifiers as rows and time points as columns.
 }

\keyword{datasets}
