\name{Data_Jena_76_metapopulations}
\alias{Data_Jena_76_metapopulations}
\docType{data}
\title{
Biomass time series data of 76 sets of populations in the Jena experiment
}
\description{
  The biomass data for all species sown within each plot are used to form 76 metapopulations. Because the number of species sown per plot ranges from 1 to 16, the number of populations within the 76 plots also ranges from 1 to 16. Each metapopulation dataset therefore includes all population-level records for its constituent species. \cr
}

\usage{data(Data_Jena_76_metapopulations)}
\format{
  \code{Data_Jena_76_metapopulations} contains 76 lists; each list is a data frame with population identifiers as rows and time points as columns.
 }

\keyword{datasets}
