% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Function.R
\name{ggiSTAY_analysis}
\alias{ggiSTAY_analysis}
\title{ggplot2 extension for plotting diversity–stability and diversity–synchrony relationships.}
\usage{
ggiSTAY_analysis(output, x_variable, by_group = NULL, model = "LMM")
}
\arguments{
\item{output}{The output returned by \code{iSTAY_Single} or \code{iSTAY_Multiple}. It must include (or be combined with) a column corresponding to the variable specified in \code{x_variable}. If \code{by_group} is not \code{NULL}, it must also include a column corresponding to the variable specified in \code{by_group}.}

\item{x_variable}{The name of the column representing the diversity variable (or another variable) to be displayed on the x-axis of the plot.}

\item{by_group}{The name of the column representing a categorical grouping variable used to color points by group. The argument is required if \code{model = "LMM"}, as the model uses it as random effect for both intercept and slope. The default is \code{NULL}.}

\item{model}{Specifies the fitting model. Use \code{model = "lm"} for a linear model; or \code{model = "LMM"} for a linear mixed model with random effects for both intercept and slope. The default is \code{model = "LMM"}.}
}
\value{
For an \code{iSTAY_Single} object, this function returns a figure showing the relationship between the diversity variable (or another variable) and stability. \cr

For an \code{iSTAY_Multiple} object, this function returns a set of plots showing the relationships between the diversity variable (or another variable) and gamma, alpha, and beta stability, together with synchrony. \cr
}
\description{
\code{ggiSTAY_analysis} is a graphical function based on the output from the functions \code{iSTAY_Single} or \code{iSTAY_Multiple}. It generates plots showing the relationships between stability (and synchrony if multiple time series are included) and an additional variable, such as diversity or another relevant factor.
}
\examples{

data("Data_Jena_20_metacommunities")
data("Data_Jena_76_community_populations")
data("Data_Jena_462_populations")
data("Data_Jena_hierarchical_structure")

## Single time series analysis
# Assessing diversity-stability relationships based on 76 individual plots 
# See Example 2 in the iSTAY vignette for the output

communities_aggregated <- do.call(rbind, Data_Jena_20_metacommunities)

output_communities_aggregated_div <- iSTAY_Single(data = communities_aggregated, 
                                   order.q = c(1,2), Alltime = TRUE)
                                   
output_communities_aggregated_div <- data.frame(output_communities_aggregated_div,
                                        log2_sowndiv = log2(as.numeric(do.call(rbind,
                                        strsplit(output_communities_aggregated_div[,1],
                                        "[._]+"))[,2])),
                                        block=do.call(rbind,
                                        strsplit(output_communities_aggregated_div[,1],
                                        "[._]+"))[,1])


ggiSTAY_analysis(output = output_communities_aggregated_div,
                 x_variable = "log2_sowndiv", 
                 by_group = "block", 
                 model = "LMM")

# Assessing diversity-stability relationships based on 462 individual populations
# See Example 4 in the iSTAY vignette for the output.

individual_populations <- Data_Jena_462_populations

output_individual_populations_div <- iSTAY_Single(data = individual_populations,
                                                 order.q = c(1,2), Alltime=TRUE)
                                                 
output_individual_populations_div <- data.frame(
                                          output_individual_populations_div,
                                          log2_sowndiv = log2(as.numeric(do.call(rbind,
                                          strsplit(output_individual_populations_div[,1],
                                          "[._]+"))[,3])),
                                          block = do.call(rbind,
                                          strsplit(output_individual_populations_div[,1],
                                          "[._]+"))[,2])

ggiSTAY_analysis(output=output_individual_populations_div,
                 x_variable="log2_sowndiv",
                 by_group="block", 
                 model="LMM")


## Multiple time series analysis
# Assessing relationships between diversity and gamma, alpha, and beta stability,
# as well as synchrony, across 76 communities
# See Example 6 in the iSTAY vignette for the output.

# Equal-weighted analysis

communities <- Data_Jena_76_community_populations

output_communities_equal_div <- iSTAY_Multiple(
                                     data = communities,
                                     order.q = c(1, 2),
                                     equal_weights = TRUE,
                                     Alltime = TRUE)

output_communities_equal_div <- data.frame(
                                     output_communities_equal_div, 
                                     log2_sowndiv = log2(as.numeric(do.call(rbind, 
                                     strsplit(output_communities_equal_div[, 1],
                                     "[._]+"))[, 3])),
                                     block = do.call(rbind, 
                                     strsplit(output_communities_equal_div[, 1]
                                     , "_"))[, 2])

ggiSTAY_analysis(output = output_communities_equal_div,
                 x_variable = "log2_sowndiv",
                 by_group = "block",
                 model = "LMM")

# Biomass-weighted analysis

output_communities_biomass_div <- iSTAY_Multiple(
                                       data = communities,
                                       order.q = c(1, 2),
                                       equal_weights = FALSE,
                                       Alltime = TRUE)
                                       
output_communities_biomass_div <- data.frame(
                                       output_communities_biomass_div,
                                       log2_sowndiv = log2(as.numeric(do.call(rbind,
                                       strsplit(output_communities_biomass_div[, 1],
                                       "[._]+"))[, 3])),
                                       block = do.call(rbind,
                                       strsplit(output_communities_biomass_div[, 1]
                                       , "_"))[, 2])

ggiSTAY_analysis(output = output_communities_biomass_div,
                 x_variable = "log2_sowndiv",
                 by_group = "block",
                 model = "LMM")


# Assessing relationships between diversity and gamma, alpha, and beta stability, 
# as well as synchrony, across 20 metacommunities
# See Example 8 in the iSTAY vignette for the output.

# Equal-weighted analysis

metacommunities <- Data_Jena_20_metacommunities

output_metacommunities_equal_div <- iSTAY_Multiple(
                                     data = metacommunities,
                                     order.q = c(1, 2),
                                     equal_weights = TRUE,
                                     Alltime = TRUE)
                                     
output_metacommunities_equal_div <- data.frame(
                                     output_metacommunities_equal_div,
                                     log2_sowndiv = log2(as.numeric(do.call(rbind,
                                     strsplit(output_metacommunities_equal_div[, 1],
                                     "_"))[, 2])),
                                     block = do.call(rbind,
                                     strsplit(output_metacommunities_equal_div[, 1],
                                     "_"))[, 1])

ggiSTAY_analysis(output = output_metacommunities_equal_div,
                 x_variable = "log2_sowndiv",
                 by_group = "block",
                 model = "LMM")

# Biomass-weighted analysis

output_metacommunities_biomass_div <- iSTAY_Multiple(
                                       data = metacommunities,
                                       order.q = c(1, 2),
                                       equal_weights = FALSE,
                                       Alltime = TRUE)
                                       
output_metacommunities_biomass_div <- data.frame(
                                       output_metacommunities_biomass_div,
                                       log2_sowndiv = log2(as.numeric(do.call(rbind,
                                       strsplit(output_metacommunities_biomass_div[, 1],
                                       "_"))[, 2])),
                                       block = do.call(rbind,
                                       strsplit(output_metacommunities_biomass_div[, 1],
                                       "_"))[, 1])

ggiSTAY_analysis(output = output_metacommunities_biomass_div,
                 x_variable = "log2_sowndiv",
                 by_group = "block",
                 model = "LMM")


}
