% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Function.R
\name{ggStay_qprofile}
\alias{ggStay_qprofile}
\title{ggplot2 extension for a Stay_Single, Stay_Multiple or Stay_Hier object with q-profile.}
\usage{
ggStay_qprofile(output)
}
\arguments{
\item{output}{the output obtained from \code{Stay_Single}, \code{Stay_Multiple} or \code{Stay_Hier}.}
}
\value{
For a \code{Stay_Single} object, this function return a figure of q-profile for stability .
For a \code{Stay_Multiple} object, this function return a figure that contains q-profile for (Gamma, Alpha, Beta) stability and synchrony.
For a \code{Stay_Hier} object, this function return a figure that contains q-profile for gamma stability of highest hierarchical level and alpha stability of other hierarchical level. And it also provides a figure about the relationship between each decomposition of overall stability and the synchrony of each hierarchical level and order.q.
}
\description{
\code{ggStay_qprofile} is a graphical function that based on the output from the function \code{Stay_Single}, \code{Stay_Multiple} or \code{Stay_Hier}. It provides to graph the q-profile of stability (and synchrony if is multiple assemblages).
}
\examples{
data("Jena_plot_biomass_data")
data("Jena_species_biomass_data")
data("Jena_hierarchical_biomass_data")
data("Jena_hierarchical_mat")

## Single assemblage
# Stability of each single plot
single_plot <- do.call(rbind, Jena_plot_biomass_data)
output_single_plot_q <- Stay_Single(data=single_plot[c(12,65),],
                                    order.q=seq(0.1,2,0.1), Alltime=TRUE)
ggStay_qprofile(output=output_single_plot_q)

# Stability of each single species
single_species <- do.call(rbind, Jena_species_biomass_data)
output_single_species_q <- Stay_Single(data=single_species[c(41,45),],
                                       order.q=seq(0.1,2,0.1), Alltime=TRUE)
ggStay_qprofile(output=output_single_species_q)


## Multiple assemblages
# Stability of multiple plots
multiple_plot <- Jena_plot_biomass_data
output_multi_plot_q <- Stay_Multiple(data=multiple_plot[c(1,13)],
                                         order.q=seq(0.1,2,0.1), Alltime=TRUE)
ggStay_qprofile(output=output_multi_plot_q)

# Stability of multiple species in plot
multiple_species <- Jena_species_biomass_data
output_multi_species_q <- Stay_Multiple(data=multiple_species[c(4,69)],
                                            order.q=seq(0.1,2,0.1), Alltime=TRUE)
ggStay_qprofile(output=output_multi_species_q)


## Hierarchies
output_hier_q <- Stay_Hier(data=Jena_hierarchical_biomass_data,
                           mat=Jena_hierarchical_mat,
                           order.q=seq(0.1,2,0.1), Alltime=TRUE)
ggStay_qprofile(output=output_hier_q)

}
