% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Function.R
\name{iSTAY_Single}
\alias{iSTAY_Single}
\title{Calculate stability for a single time series.}
\usage{
iSTAY_Single(
  data,
  order.q = c(1, 2),
  Alltime = TRUE,
  start_T = NULL,
  end_T = NULL
)
}
\arguments{
\item{data}{A \code{vector} of time series data, or a \code{data.frame} with sampling units as rows and time points as columns.}

\item{order.q}{A numerical vector specifying the orders of stability. Default is c(1,2).}

\item{Alltime}{Logical (\code{TRUE} or \code{FALSE}), indicating whether to use all time points in the data.}

\item{start_T}{(Applicable only if \code{Alltime = FALSE}) a positive integer specifying the starting column (time point) for the analysis interval.}

\item{end_T}{(Applicable only if \code{Alltime = FALSE}) a positive integer specifying the ending column (time point) for the analysis interval.}
}
\value{
a data frame with the following columns: \cr
Dataset: the input dataset \cr
Order_q: order of stability \cr
Stability: stability measures of order q
}
\description{
\code{iSTAY_Single} computes the stability of order q for a single time series.
}
\examples{
# Computing stability of an individual community

data("Data_Jena_20_metacommunities")

community_aggregated <- do.call(rbind, Data_Jena_20_metacommunities)

output_community_aggregated <- iSTAY_Single(data = community_aggregated, 
                                        order.q=c(1,2), Alltime = TRUE)
                                        
output_community_aggregated

# Computing stability of an individual population

data("Data_Jena_462_populations")

individual_populations <- Data_Jena_462_populations

output_individual_populations <- iSTAY_Single(data = individual_populations, 
                                              order.q = c(1,2), Alltime = TRUE)
                                              
output_individual_populations

}
