\name{Jena_hierarchical_mat}
\alias{Jena_hierarchical_mat}
\docType{data}
\title{
Hierarchical structure matrix corresponding to \code{Jena_hierarchical_data}
}
\description{
  This dataset includes three columns which are "block", "plot" and "species" the hirarchical level of Jena experiment. Each row of this data is corresponding to the same row in \code{Jena_hierarchical_data}, and record the names of block, plot and species that the row belongs to.
}
\usage{data(Jena_hierarchical_mat)}
\format{
  \code{Jena_hierarchical_mat} is a data.frame with "block", "plot" and "species" three columns and 462 rows which is sames as the number of rows in \code{Jena_hierarchical_data}.
 }
\references{
Wagg, C., Roscher, C., Weigelt, A. et al. (2022) Biodiversity–stability relationships strengthen over time in a long-term grassland experiment. Nature Communications 13, 7752. \cr\cr
}

\keyword{datasets}
