\name{Jena_plot_biomass_data}
\alias{Jena_plot_biomass_data}
\docType{data}
\title{
17-years biomass data of plots in Jena experiment
}
\description{
  In Jena experiment, there are 4 blocks (B1, B2, B3, B4) with a total of 76 plots. For each plot 1, 2, 4, 8 or 16 plants are sown. There are approximately 3 or 4 plots in each block for each number of plant species planted. The set includes 20 lists that are combinations of four blocks and number of five different plant species planted. In each list, it is a plots/assemblages by time dataframe. For example, the list "B1_16" is a dataframe that contains the 22-year biomass of all plants in the plots, which are in block B1 and has been sown with 16 plant species.\cr
}

\usage{data(Jena_plot_biomass_data)}
\format{
  \code{Jena_plot_biomass_data} contains 20 lists. Each list is a dataframe that the rows are plots/assemblages and the columns are the time. For each row, it record the biomass  of the plot/assemblage during the time.
 }

\references{
Wagg, C., Roscher, C., Weigelt, A. et al. (2022) Biodiversity–stability relationships strengthen over time in a long-term grassland experiment. Nature Communications 13, 7752. \cr\cr
}

\keyword{datasets}
