% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Function.R
\name{Stab_Single}
\alias{Stab_Single}
\title{Calculate stability of the time series data for single assemblage.}
\usage{
Stab_Single(
  data,
  order.q = c(1, 2),
  Alltime = TRUE,
  start_T = NULL,
  end_T = NULL
)
}
\arguments{
\item{data}{can be input as a \code{vector} of time series data, or \code{data.frame/matrix} (assemblages by times).}

\item{order.q}{a numerical vector specifying the orders of stability. Default is c(1,2).}

\item{Alltime}{\code{TRUE} or \code{FALSE}, to decide whether to use all the times in the data.}

\item{start_T}{(argument only for \code{Alltime = FALSE}) a positive integer specifying the start column of time in the data.}

\item{end_T}{(argument only for \code{Alltime = FALSE}) a positive integer specifying the end column of time in the data.}
}
\value{
a dataframe with columns "Assemblage", "Order_q" and "Stability".
}
\description{
\code{Stab_Single} is a function that calculate stability of the time series data (like biomass, productivity, etc.) for single assemblage.
}
\examples{
data("Jena_experiment_plant_data")
single_data <- do.call(rbind, Jena_experiment_plant_data)
output <- Stab_Single(data=single_data,
                      order.q=c(1,2), Alltime=TRUE)
output

}
