% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Function.R
\name{Stab_Syn_ggplot}
\alias{Stab_Syn_ggplot}
\title{Calculate stability and synchrony of the time series data for multiple assemblages.}
\usage{
Stab_Syn_ggplot(output)
}
\arguments{
\item{output}{the output obtained from \code{Stab_Single} or \code{Stab_Syn_Multiple}.}
}
\value{
For an \code{Stab_Single} object, this function return a figure of order of q vs. stability.
For an \code{Stab_Syn_Multiple} object, this function return two figures that are order of q vs. (Gamma, Alpha, Beta) stability and order of q vs. synchrony.
}
\description{
\code{Stab_Syn_ggplot} is a function that calculate (Gamma, Alpha and Beta) stability and synchrony of the time series data (like biomass, productivity, etc.) for multiple assemblages.
}
\examples{
data("Jena_experiment_plant_data")

output_single <- Stab_Single(data=Jena_experiment_plant_data[[1]],
                             order.q=seq(0.2,2,0.2), Alltime=TRUE)
Stab_Syn_ggplot(output_single)

output_multi <- Stab_Syn_Multiple(Jena_experiment_plant_data,
                                  order.q=seq(0.2,2,0.2), Alltime=TRUE)
Stab_Syn_ggplot(output_multi)

}
