% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Function.R
\name{Stab_Syn_Multiple}
\alias{Stab_Syn_Multiple}
\title{Calculate stability and synchrony of the time series data for multiple assemblages.}
\usage{
Stab_Syn_Multiple(
  data,
  order.q = c(1, 2),
  Alltime = TRUE,
  start_T = NULL,
  end_T = NULL
)
}
\arguments{
\item{data}{can be input as a \code{data.frame/matrix} (assemblages by times), or a \code{list} of \code{data.frames} with each dataframe representing a assemblages-by-times data.}

\item{order.q}{a numerical vector specifying the orders of stability and synchrony. Default is c(1,2).}

\item{Alltime}{\code{TRUE} or \code{FALSE}, to decide whether to use all the times in (every) dataframe.}

\item{start_T}{(argument only for \code{Alltime = FALSE}) a positive integer specifying the start column of time in (every) dataframe.}

\item{end_T}{(argument only for \code{Alltime = FALSE}) a positive integer specifying the end column of time in (every) dataframe.}
}
\value{
a dataframe with columns "Place", "Order_q", "Gamma", "Alpha", "Beta" and "Synchrony".
}
\description{
\code{Stab_Syn_Multiple} is a function that calculate (Gamma, Alpha and Beta) stability and synchrony of the time series data (like biomass, productivity, etc.) for multiple assemblages.
}
\examples{
data("Jena_experiment_plant_data")
multiple_data <- Jena_experiment_plant_data
output_multi <- Stab_Syn_Multiple(data=multiple_data, order.q=c(1,2), Alltime=TRUE)
output_multi


}
