% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Function.R
\name{Stay_Hier}
\alias{Stay_Hier}
\title{Calculate stability of the time series data for hierarchical structure.}
\usage{
Stay_Hier(
  data,
  mat,
  order.q = c(1, 2),
  Alltime = TRUE,
  start_T = NULL,
  end_T = NULL
)
}
\arguments{
\item{data}{can be input as \code{data.frame} (assemblages by times).}

\item{mat}{hierarchical structure of data.}

\item{order.q}{a numerical vector specifying the orders of stability. Default is c(1,2).}

\item{Alltime}{\code{TRUE} or \code{FALSE}, to decide whether to use all the times in the data.}

\item{start_T}{(argument only for \code{Alltime = FALSE}) a positive integer specifying the start column of time in the data.}

\item{end_T}{(argument only for \code{Alltime = FALSE}) a positive integer specifying the end column of time in the data.}
}
\value{
a dataframe with columns "Hier", "Order_q", and stability "Gamma", "Alpha", "Beta" and "Synchrony".
}
\description{
\code{Stay_Hier} is a function that calculate stability of the time series data (like biomass, productivity, etc.) for hierarchical structure.
}
\examples{

data("Jena_hierarchical_biomass_data")
data("Jena_hierarchical_mat")
output_hier <- Stay_Hier(data=Jena_hierarchical_biomass_data, mat=Jena_hierarchical_mat,
                         order.q=c(1,2), Alltime=TRUE)
output_hier


}
