% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Function.R
\name{Stay_Multiple}
\alias{Stay_Multiple}
\title{Calculate stability and synchrony of time-series data for multiple assemblages.}
\usage{
Stay_Multiple(
  data,
  order.q = c(1, 2),
  Alltime = TRUE,
  start_T = NULL,
  end_T = NULL
)
}
\arguments{
\item{data}{can be input as a \code{data.frame/matrix} (assemblages by times), or a \code{list} of \code{data.frames} with each dataframe representing a assemblages-by-times data.}

\item{order.q}{a numerical vector specifying the orders of stability and synchrony. Default is c(1,2).}

\item{Alltime}{\code{TRUE} or \code{FALSE}, to decide whether to use all the times in (every) dataframe.}

\item{start_T}{(argument only for \code{Alltime = FALSE}) a positive integer specifying the start column of time in (every) dataframe.}

\item{end_T}{(argument only for \code{Alltime = FALSE}) a positive integer specifying the end column of time in (every) dataframe.}
}
\value{
a dataframe with columns "Site", "Order_q", "Gamma", "Alpha", "Beta" and "Synchrony".
}
\description{
\code{Stay_Multiple} is a function that computes gamma, alpha, and beta stability, as well as synchrony, for time-series data (e.g., biomass, productivity) across multiple assemblages.
}
\examples{
# Stability of multiple plots
data("Jena_plot_biomass_data")
multiple_plot <- Jena_plot_biomass_data
output_multi_plot <- Stay_Multiple(data=multiple_plot, order.q=c(1,2), Alltime=TRUE)
output_multi_plot

# Stability of multiple species in each plot
data("Jena_species_biomass_data")
multiple_species <- Jena_species_biomass_data
output_multi_species <- Stay_Multiple(data=multiple_species, order.q=c(1,2), Alltime=TRUE)
output_multi_species


}
