\name{Data_Jena_20_metacommunities}
\alias{Data_Jena_20_metacommunities}
\docType{data}
\title{
Biomass time series data of 20 sets of plots in the Jena experiment
}
\description{
  Because the Jena Experiment includes four blocks and five richness levels, all 20 block–richness combinations are considered, resulting in 20 sets of plots. In other words, all 76 plots are grouped into 20 sets based on their block identifiers and species richness levels. Each set is treated as a metacommunity, comprising three or four plots (communities). All plots within a metacommunity share the same number of species but differ in species composition. \cr
}

\usage{data(Data_Jena_20_metacommunities)}
\format{
  \code{Data_Jena_20_metacommunities} contains 20 lists; each list is a data frame with plot identifiers as rows and time points as columns.
 }


\keyword{datasets}
