\name{Data_Jena_462_populations}
\alias{Data_Jena_462_populations}
\docType{data}
\title{
Biomass time series data of 462 populations in the Jena experiment
}
\description{
  In the Jena experiment, the biomass of each species sown in a plot was recorded annually from 2003 to 2024, except in 2004. This dataset includes the biomass time series of all 462 populations across 76 plots.
}
\usage{data(Data_Jena_462_populations)}
\format{
  \code{Data_Jena_462_populations} is a population-by-time (462 × 21) data frame with population identifiers as rows and time points as columns.
 }

\keyword{datasets}
