\name{Data_Jena_76_communities}
\alias{Data_Jena_76_communities}
\docType{data}
\title{
Biomass time series data of 76 communities in the Jena experiment
}
\description{
  The biomass data for all species sown within each plot are used to form 76 community datasets. Because the number of species sown per plot ranges from 1 to 16, the number of populations within each of the 76 communities also ranges from 1 to 16. Each community dataset therefore includes all population-level records for its constituent species. \cr
}

\usage{data(Data_Jena_76_communities)}
\format{
  \code{Data_Jena_76_communities} contains 76 lists; each list is a data frame with plot identifiers as rows and time points as columns.
 }

\keyword{datasets}
