% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Function.R
\name{ggStab_Syn_qprofile}
\alias{ggStab_Syn_qprofile}
\title{ggplot2 extension for a Stab_Single or Stab_Syn_Multiple object with q-profile.}
\usage{
ggStab_Syn_qprofile(output)
}
\arguments{
\item{output}{the output obtained from \code{Stab_Single} or \code{Stab_Syn_Multiple}.}
}
\value{
For an \code{Stab_Single} object, this function return a figure of q-profile for stability .
For an \code{Stab_Syn_Multiple} object, this function return two figures which are q-profile for (Gamma, Alpha, Beta) stability and q-profile for synchrony.
}
\description{
\code{ggStab_Syn_qprofile} is a graphical function that based on the output from the function \code{Stab_Single} or \code{Stab_Syn_Multiple}. It provides to graph the q-profile of stability (and synchrony if is multiple assemblages).
}
\examples{
data("Jena_experiment_plant_data")

## Single assemblage
single_data <- do.call(rbind, Jena_experiment_plant_data)
output_single <- Stab_Single(data=single_data[1:4,], order.q=seq(0.1,2,0.05), Alltime=TRUE)
ggStab_Syn_qprofile(output=output_single)


## Multiple assemblages
output_multi <- Stab_Syn_Multiple(Jena_experiment_plant_data[1:4],
                                  order.q=seq(0.1,2,0.05), Alltime=TRUE)
ggStab_Syn_qprofile(output=output_multi)

}
